Reverted adminUpdate to previous state, added adminSearch with changes previously added to adminUpdate
